Generate .xcodeproj for Swift package if it doesn't include one
Many projects that support Swift Package Manager include an .xcodeproj in their repo, but for those that don't it would be nice to use them in TestDrive just the same. This change takes effect when a cloned repo doesn't include an .xcodeproj but does include a Package.swift file. In that case it will use `swift package generate-xcodeproj` to create an .xcodeproj in a known location automatically.